+++ lib/features/home/presentation/views/constant.dart
@@ -1,0 +1,9 @@
+abstract class HomeConstant{
+
+static List<String> sectionsName = [
+  'Vegetables',
+  'Fruits',
+  'Dry Fruits',
+];
+}
+
