+++ lib/features/home/presentation/views/widgets/page_view_element.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+import 'custom_section_widget.dart';
+
+class PageViewElement extends StatelessWidget {
+  const PageViewElement({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return const SingleChildScrollView(
+      child: Column(
+        children: [
+          CustomSectionWidget(),
+          CustomSectionWidget(),
+          CustomSectionWidget(),
+        ],
+      ),
+    );
+  }
+}
